final touch on the server setup

+++ backend/controllers/product.controller.js
@@ -33,29 +33,34 @@ export const createProduct = async (req, res) => {
 
 export const updateProduct = async (req, res) => {
   const { id } = req.params;
+
   const product = req.body;
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
     return res
       .status(404)
-      .json({ success: false, message: "Invalid product id" });
+      .json({ success: false, message: "Invalid Product Id" });
   }
 
   try {
-    const updateProduct = await Product.findByIdAndUpdate(id, product, {
+    const updatedProduct = await Product.findByIdAndUpdate(id, product, {
       new: true,
     });
-    res
-      .status(200)
-      .json({ success: true, message: "Product updated successfully" });
+    res.status(200).json({ success: true, data: updatedProduct });
   } catch (error) {
-    console.log("error in updating product:", error.message);
     res.status(500).json({ success: false, message: "Server Error" });
   }
 };
 
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
+
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Invalid Product Id" });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     res
@@ -63,6 +68,6 @@ export const deleteProduct = async (req, res) => {
       .json({ success: true, message: "Product deleted successfully" });
   } catch (error) {
     console.log("error in deleting product:", error.message);
-    res.status(404).json({ success: false, message: "Product not found." });
+    res.status(500).json({ success: false, message: "Server Error" });
   }
 };
